backend-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
faiss-cpu==1.8.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy>=1.24.0

